requirements.txt
pdfplumber>=0.9.0
PyPDF2>=3.0.0
python-docx>=0.8.11
ebooklib>=0.18
beautifulsoup4>=4.12.0
lxml>=4.9.0
pytesseract>=0.3.10
Pillow>=10.0.0
pandas>=2.0.0
pyarrow>=12.0.0
pyyaml>=6.0
click>=8.1.0
nltk>=3.8
spacy>=3.6.0
scikit-learn>=1.3.0
tqdm>=4.65.0
chardet>=5.0.0
unicodedata2>=15.0.0
streamlit>=1.28.0
plotly>=5.15.0
streamlit-aggrid>=0.3.4
streamlit-option-menu>=0.3.6